pyproject.toml
[tool.poetry]
name = "growth-dash"
version = "0.1.0"
description = "Dashboard of growth statistics in infants with clefts (built with Streamlit)"
homepage = ""
repository = "https://github.com/jonjlee/growth-dash"
authors = ["Jonathan Lee <jonathan.lee@seattlechildrens.org>"]
keywords = ["cleft", "craniofacial", "pediatrics", "streamlit"]

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.1.2"
PyNaCl = "^1.4.0"
pandas = "^1.2.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
streamlit = "^0.76.0"

[tool.poetry.scripts]
growth-dash = "growth_dash.console:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
